+++ CHANGELOG.md
@@ -36,3 +36,10 @@ Refer to README for any information.
 * Improved behavior of `readTags` method.
 * Migrated [`AudiotaggerPlugin.java`](https://github.com/Samurai016/Audiotagger/blob/master/android/src/main/java/com/nicolorebaioli/audiotagger/AudiotaggerPlugin.java) to [1.12 Flutter plugin APIs](https://flutter.dev/docs/development/packages-and-plugins/plugin-api-migration).
 * Improved testing files.
+
+## 2.2.0
+
+* Added `AudioFile` class (solved [issue #14](https://github.com/Samurai016/Audiotagger/issues/14)).
+* Added `readAudioFile` and `readAudioFileAsMap` methods (solved [issue #14](https://github.com/Samurai016/Audiotagger/issues/14)).
+* Fixed small bugs.
+* Updated documentation.